Add Config Reader FileWatching tests
- Refactor the Reader slightly to allow adjusting changeFlushPeriod
and changeDelayPeriod from tests.
- Move the original black box config tests to a package test file.
